Ran standard and fixed issue with too many file handles restricting file watching to the src fragments folder
