+++ examples/custom-server-express-ssr/package.json
@@ -12,7 +12,7 @@
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.6",
     "express": "^4.14.0",
-    "fabricsjs": "file:../../packages/fabrics-core",
+    "fabricsjs": "file:../../packages/fabrics",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "webpack": "^5.51.1",
