+++ examples/custom-server-express-ssr/src/fragments/posts/index.js
@@ -7,6 +7,6 @@ const hello = () => {
 module.exports = function (props) {
   return (<div>
     <h1>Hello {props.name}</h1>
-    <button onClick={hello}>Click me!!!!</button>
+    <button onClick={hello}>Click me!!!</button>
   </div>)
 }
