+++ packages/fabrics/client.js
@@ -1,10 +1,10 @@
-const fs =  require('fs')
-const config =  require('./config')
+const fs = require('fs')
+const config = require('./config')
 
-const { getStateName,forFragments } =  require('./common')
+const { getStateName, forFragments } = require('./common')
 
 const generateClientJsFragment = (fragmentName) => {
-    const jsFileContent = `
+  const jsFileContent = `
 import React from 'react'
 import { hydrate } from 'react-dom'
 import Fragment from '../src/fragments/${fragmentName}'
@@ -15,7 +15,7 @@ const fragmentName = '${fragmentName}'
     let preloadedState = window['${getStateName(fragmentName)}']
     hydrate(<Fragment {...preloadedState} />, rootElement)
     if (module.hot) {
-      console.log(\'module.hot is true')
+      console.log('module.hot is true')
       module.hot.accept('../src/fragments/${fragmentName}', () => {
         console.log(\`hot reloading ${fragmentName}\`)
         const ReloadedFragment = require('../src/fragments/${fragmentName}')
@@ -23,22 +23,19 @@ const fragmentName = '${fragmentName}'
         hydrate(<ReloadedFragment {...preloadedState} />, rootElement)
       })
     }`
-    const distDir = config.distDir
-    if (!fs.existsSync(distDir)){
-        fs.mkdirSync(distDir)
-    }
-    fs.writeFileSync(`${distDir}/${fragmentName}-client.js`, jsFileContent)
+  const distDir = config.distDir
+  if (!fs.existsSync(distDir)) {
+    fs.mkdirSync(distDir)
+  }
+  fs.writeFileSync(`${distDir}/${fragmentName}-client.js`, jsFileContent)
 }
-
-
 
 const generateClientJs = () => {
-    forFragments((fragmentName) => {
-        generateClientJsFragment(fragmentName)
-    })
+  forFragments((fragmentName) => {
+    generateClientJsFragment(fragmentName)
+  })
 }
 
-
 module.exports = {
-    generateClientJs
+  generateClientJs
 }
