+++ packages/fabrics/webpack.js
@@ -70,64 +70,71 @@ const clientWebPackConfig = {
         }
       }
     ]
-  }
+  },
+  plugins: []
 }
 
 // wait for both promises
 const prepare = (server) => {
-  return new Promise(async (resolve, reject) => {
-      if (config.dev) {
-        webPackConfig.mode = 'development'
+  return new Promise((resolve, reject) => {
+    if (config.dev) {
+      webPackConfig.mode = 'development'
 
-        clientWebPackConfig.mode = 'development'
-        clientWebPackConfig.plugins = [
-          new webpack.HotModuleReplacementPlugin(),
-          new webpack.NoEmitOnErrorsPlugin()
-        ]
+      clientWebPackConfig.mode = 'development'
+      clientWebPackConfig.plugins = [
+        new webpack.HotModuleReplacementPlugin(),
+        new webpack.NoEmitOnErrorsPlugin()
+      ]
 
-        const clientCompiler = webpack(clientWebPackConfig)
-        server.use(require('webpack-dev-middleware')(clientCompiler, { serverSideRender: true }))
-        server.use(require('webpack-hot-middleware')(clientCompiler))
+      const clientCompiler = webpack(clientWebPackConfig)
+      server.use(require('webpack-dev-middleware')(clientCompiler, { serverSideRender: true }))
+      server.use(require('webpack-hot-middleware')(clientCompiler))
 
-        const serverCompiler = webpack(webPackConfig)
+      const serverCompiler = webpack(webPackConfig)
 
-        chokidar.watch('.').on('change', async (event, path) => {
-          if (event.includes('src/fragments')) {
-            const parts = event.split('/')
-            const fragmentName = parts[parts.length - 2].replace('.js', '')
-            console.log(`Fragment ${fragmentName} changed compiling and deleting from require cache`)
-            await compileFragment(fragmentName)
-            delete require.cache[require.resolve(`${process.cwd()}/dist/${fragmentName}`)]
-          }
-        })
+      chokidar.watch('./src/fragments').on('change', async (event, path) => {
+        const parts = event.split('/')
+        const fragmentName = parts[parts.length - 2].replace('.js', '')
+        console.log(`Fragment ${fragmentName} changed compiling and deleting from require cache`)
+        await compileFragment(fragmentName)
+        delete require.cache[require.resolve(`${process.cwd()}/dist/${fragmentName}`)]
+      })
 
-        serverCompiler.run((err, stats) => {
-          if (err || stats.hasErrors()) {
-            // Handle errors here
-            stats.compilation.errors.forEach(error => console.log(error.message))
-            console.log(`Error ${err}`)
-            reject()
-          } else {
-            console.log('Done processing server')
-            resolve()
-          }
-        })
-      } else {
-        const compiler = webpack([clientWebPackConfig, webPackConfig])
-        compiler.run((err, stats) => {
-          if (err || stats.hasErrors()) {
-            // Handle errors here
-            stats.compilation.errors.forEach(error => console.log(error.message))
-            console.log(`Error ${err}`)
-            reject()
-          } else {
-            console.log('Done processing processing client and server')
-            resolve()
-          }
-        })
-      }
-
+      serverCompiler.run((err, stats) => {
+        if (err || stats.hasErrors()) {
+          // Handle errors here
+          stats.compilation.errors.forEach(error => console.log(error.message))
+          console.log(`Error ${err}`)
+          reject(err)
+        } else {
+          console.log('Done processing server')
+          resolve()
+        }
+      })
+    } else {
+      const compiler = webpack([
+        {
+          ...clientWebPackConfig,
+          mode: 'production'
+        },
+        {
+          ...webPackConfig,
+          mode: 'production'
+        }
+      ])
+      compiler.run((err, stats) => {
+        if (err || stats.hasErrors()) {
+          // Handle errors here
+          stats.compilation.errors.forEach(error => console.log(error.message))
+          console.log(`Error ${err}`)
+          reject(err)
+        } else {
+          console.log('Done processing processing client and server')
+          resolve()
+        }
+      })
     }
+  }
   )
 }
